Created an HTML index and linked javascript file
Created an HTML index file with a linked javascript file that includes code for an if else statement, the use of a global variable, and the use of a local variable. The code for the local variable also includes a purposeful error to use as an example of debugging from within the dev tools of Chrome. Comments are also used throughout the javascript file.